+++ src/Veldrid/D3D12/D3D12Framebuffer.cs
@@ -51,7 +51,7 @@ namespace Veldrid.D3D12
                     if (d3dDepthTarget.SampleCount == TextureSampleCount.Count1)
                     {
                         dsvDesc.ViewDimension = DepthStencilViewDimension.Texture2D;
-                        dsvDesc.Texture2D.MipSlice = (int)description.DepthTarget.Value.MipLevel;
+                        dsvDesc.Texture2D.MipSlice = description.DepthTarget.Value.MipLevel;
                     }
                     else
                     {
@@ -63,14 +63,14 @@ namespace Veldrid.D3D12
                     if (d3dDepthTarget.SampleCount == TextureSampleCount.Count1)
                     {
                         dsvDesc.ViewDimension = DepthStencilViewDimension.Texture2DArray;
-                        dsvDesc.Texture2DArray.FirstArraySlice = (int)description.DepthTarget.Value.ArrayLayer;
+                        dsvDesc.Texture2DArray.FirstArraySlice = description.DepthTarget.Value.ArrayLayer;
                         dsvDesc.Texture2DArray.ArraySize = 1;
-                        dsvDesc.Texture2DArray.MipSlice = (int)description.DepthTarget.Value.MipLevel;
+                        dsvDesc.Texture2DArray.MipSlice = description.DepthTarget.Value.MipLevel;
                     }
                     else
                     {
                         dsvDesc.ViewDimension = DepthStencilViewDimension.Texture2DMultisampledArray;
-                        dsvDesc.Texture2DMSArray.FirstArraySlice = (int)description.DepthTarget.Value.ArrayLayer;
+                        dsvDesc.Texture2DMSArray.FirstArraySlice = description.DepthTarget.Value.ArrayLayer;
                         dsvDesc.Texture2DMSArray.ArraySize = 1;
                     }
                 }
@@ -101,14 +101,14 @@ namespace Veldrid.D3D12
                         {
                             rtvDesc.ViewDimension = RenderTargetViewDimension.Texture2DArray;
                             rtvDesc.Texture2DArray.ArraySize = 1;
-                            rtvDesc.Texture2DArray.FirstArraySlice = (int)description.ColorTargets[i].ArrayLayer;
-                            rtvDesc.Texture2DArray.MipSlice = (int)description.ColorTargets[i].MipLevel;
+                            rtvDesc.Texture2DArray.FirstArraySlice = description.ColorTargets[i].ArrayLayer;
+                            rtvDesc.Texture2DArray.MipSlice = description.ColorTargets[i].MipLevel;
                         }
                         else
                         {
                             rtvDesc.ViewDimension = RenderTargetViewDimension.Texture2DMultisampledArray;
                             rtvDesc.Texture2DMSArray.ArraySize = 1;
-                            rtvDesc.Texture2DMSArray.FirstArraySlice = (int)description.ColorTargets[i].ArrayLayer;
+                            rtvDesc.Texture2DMSArray.FirstArraySlice = description.ColorTargets[i].ArrayLayer;
                         }
                     }
                     else
@@ -116,7 +116,7 @@ namespace Veldrid.D3D12
                         if (d3dColorTarget.SampleCount == TextureSampleCount.Count1)
                         {
                             rtvDesc.ViewDimension = RenderTargetViewDimension.Texture2D;
-                            rtvDesc.Texture2D.MipSlice = (int)description.ColorTargets[i].MipLevel;
+                            rtvDesc.Texture2D.MipSlice = description.ColorTargets[i].MipLevel;
                         }
                         else
                         {
